HspInterpreter

Documentation

data LangRule

Constructors

MkLangRule

Fields

token :: String

phones :: [Phone]

type Phone = String

langRules

Arguments

:: String
ID of language - for example "pol"

-> IO [LangRule]

List of langRules found in .hsp file of language with given ID

strToLangRule

Arguments

:: String
String containing a LangRule - for example "ni -> ni,i"

-> LangRule

Helper method used to turn contents of .hsp file into langrules

lines'

Arguments

:: String
String containing newLine characters

-> [String]

Just like embedded lines method, but omitting empty lines

splitStr

Arguments

:: String
String to split

-> String
Regular expression to check for and split on

-> [String]

Splitting String into a list based on a regex

despace

Arguments

:: String
String to filter

-> String

Filters all spaces out of a given String

sortLangRules

Arguments

:: [LangRule]
List of langRules to sort

-> [LangRule]

Sorts a list langRules according to length of their tokens firstly and alphabetically secondarily

sortAlph

Arguments

:: [LangRule]
List of langRules to sort

-> [LangRule]

Sorts a list of langRules alphabetically according to their tokens

data Alias

Constructors

MkAlias

Fields

alias :: String

matches :: [Phone]

strToAlias

Arguments

:: String
String containing an Alias - for example "vow -> a,ą,e,ę,i,o,ó,u,y"

-> Alias

Makes an alias out of a String from .hsp file

aliases

Arguments

:: String
ID of language - for exaple "pol"

-> IO [Alias]

List of aliases from given language (.hsp file)

data AliasRule

Constructors

MkAliasRule

Fields

regex :: [Phone]

output :: [Phone]

aliasRulesAliased

Arguments

:: String
ID of language - for exaple "pol"

-> IO [AliasRule]

List of aliasRules from given language (.hsp file)

strToAliasRule

Arguments

:: String
String containing an aliasRule - for example "hardcon,i,vow -> $0,j,$1"

-> AliasRule

Makes an aliasRule out of a String from .hsp file

unAliasRule

Arguments

:: [Alias]
List of aliases to consider

-> AliasRule
aliasRule to unAlias

-> [AliasRule]

Generates a list of aliasRules from the given unaliased aliasRule: example:

unAliasRule [{alias="vovel"}, matches=["a,e,i,o,u,y"]] {regex=[vovel,"i"], output=["$0","j"]} 

would produce:

[{regex=["a","i"], output=["a","j"]},

{regex=["e","i"], output=["e","j"]},

{regex=["i","i"], output=["i","j"]},

{regex=["o","i"], output=["o","j"]},

{regex=["u","i"], output=["u","j"]},

{regex=["y","i"], output=["y","j"]}]

aliasRules

Arguments

:: String
ID of language - for exaple "pol"

-> IO [AliasRule]

Returns a list of unaliased rules from given language (.hsp file)

sortAliasRules

Arguments

:: [AliasRule]
List of aliasRules to sort

-> [AliasRule]

Sorts a list of aliasRules according to length of regex

replaceOccurences

Arguments

:: Eq a

=> a
Element of list to replace

-> [a]
List to replace regex with

-> [a]
List to make replacements in

-> [a]

Replaces all of occurences of an element of a list with a list of other elements