Phonemizer

Documentation

phonemize

Arguments

:: String
ID of language - for example "p

-> String
String to translate into words

-> IO [[Phone]]

Given the ID of language and a String returns a list of "words" - lists of phones

matchLangRule

Arguments

:: String
String to match the token of langRule to

-> [LangRule]
List of available langRules to match to

-> LangRule
Matched langRule

Given a String and a list of LangRules returns the LangRule whose token is the same as the beginning of the String

considerAliases

Arguments

:: [AliasRule]
List of aliasRules to consider

-> [Phone]
Word to modify based on aliasRules

-> [Phone]

Modifies the list of phones based on AliasRules

matchAliasRule

Arguments

:: [Phone]
List of phones to match the token of aliasRule to

-> [AliasRule]
List of available aliasRules to consider

-> Maybe AliasRule
Matched langRule

Given a word (a list of phones) and a list of AliasRules returns the AliasRule whose regex is the same as the beginning of the word