Soundgluer

Documentation

waveExtension :: FilePath

Extension of audio files representing phonems.

pathSeparator :: FilePath

Probably a naive reimplementation of some builtin.

langsDirectory :: FilePath

Name of the root languages directory.

stdLang :: FilePath

Name of the folder containing standard phonems (like silence), which appear accross all languages.

waveHeaderPath :: FilePath

Path of the wave header used during generation of the output.

glueSpeech

Arguments

:: String
Name of the language. It should match the name of the language folder in the langsDirectory.

-> [[Phone]]
List of phonemized words

-> String
Path of the output file

-> IO ()

Converts a list of phonemized words in some language into an audio representation and writes it to a file.

loadLangAudio

Arguments

:: String
Language name with a matching folder in langsDirectory

-> IO (Map Phone Builder)
Map from phonem name to its audio data as a lazy ByteString Builder

Loads lazily phonems of a given language into memory.

isWave

Arguments

:: FilePath
Path of the checked file

-> Bool
Is it a WAV?

Checks using extension if file has the WAV format.

phoneName

Arguments

:: FilePath
File name

-> String
Extracted phone name

Extracts the phone name from a file name.

getLangPath

Arguments

:: String
Name of the language matching its folder name

-> FilePath
Path to a specific language folder

Returns path to the directory of a specified language.

getAudioData

Arguments

:: FilePath
Language directory as returned by getLangPath

-> FilePath
Name of a wave file to read

-> IO Builder
Lazy ByteString Builder

Loads into memory raw audio data without header from a single wave file and returns a Builder of a lazy ByteString.

generateHeader :: IO ()

Utility for generating a wave file containing only a header and no audio data.